move appender to an inner class
